test(api-core): move firmware upgrade coverage to integration suites
