+++ crates/api-core/src/tests/host_bmc_firmware_test.rs
@@ -21,9 +21,7 @@ use std::time::Duration;
 
 use carbide_firmware::test_support::script_setup;
 use carbide_machine_controller::config::{FirmwareGlobal, TimePeriod};
-use carbide_machine_controller::handler::{
-    MAX_FIRMWARE_UPGRADE_RETRIES, MAX_NEW_FIRMWARE_REPORTED_RESET_RETRIES,
-};
+use carbide_machine_controller::handler::MAX_FIRMWARE_UPGRADE_RETRIES;
 use common::api_fixtures::instance::TestInstance;
 use common::api_fixtures::{
     self, TestEnv, TestManagedHost, create_test_env_with_overrides, get_config,
@@ -1808,633 +1806,6 @@ async fn test_manual_firmware_upgrade_workflow(pool: sqlx::PgPool) -> CarbideRes
     let host = mh.host().db_machine(&mut txn).await;
 
     assert!(host.manual_firmware_upgrade_completed.is_none());
-
-    Ok(())
-}
-
-#[crate::sqlx_test]
-async fn test_report_scout_firmware_upgrade_status(pool: sqlx::PgPool) -> CarbideResult<()> {
-    const UPGRADE_TASK_ID: &str = "scout-upgrade-task-id";
-
-    let env = create_test_env(pool).await;
-    let mh = common::api_fixtures::create_managed_host(&env).await;
-
-    // Manually put the machine into WaitingForScoutUpgrade state
-    let mut txn = env.pool.begin().await.unwrap();
-    let host = mh.host().db_machine(&mut txn).await;
-    let waiting_state = ManagedHostState::HostReprovision {
-        reprovision_state: HostReprovisionState::WaitingForScoutUpgrade {
-            upgrade_task_id: UPGRADE_TASK_ID.to_string(),
-            firmware_type: FirmwareComponentType::Bmc,
-            final_version: "1.2.3".to_string(),
-            power_drains_needed: None,
-            started_at: chrono::Utc::now(),
-            deadline: chrono::Utc::now() + chrono::TimeDelta::minutes(60),
-            task_json: String::new(),
-            result: None,
-        },
-        retry_count: 0,
-    };
-    db::machine::advance(&host, &mut txn, &waiting_state, None)
-        .await
-        .unwrap();
-    txn.commit().await.unwrap();
-
-    // Call the RPC endpoint with a successful result
-    env.api
-        .report_scout_firmware_upgrade_status(Request::new(
-            rpc::forge::ScoutFirmwareUpgradeStatusRequest {
-                machine_id: Some(mh.host().id),
-                success: true,
-                exit_code: 0,
-                stdout: "upgrade complete".to_string(),
-                stderr: String::new(),
-                error: String::new(),
-                upgrade_task_id: UPGRADE_TASK_ID.to_string(),
-            },
-        ))
-        .await
-        .unwrap();
-
-    // Verify the result was stored
-    let mut txn = env.pool.begin().await.unwrap();
-    let host = mh.host().db_machine(&mut txn).await;
-    let ManagedHostState::HostReprovision {
-        reprovision_state, ..
-    } = host.current_state()
-    else {
-        panic!("Not in HostReprovision");
-    };
-    let HostReprovisionState::WaitingForScoutUpgrade { result, .. } = reprovision_state else {
-        panic!("Not in WaitingForScoutUpgrade");
-    };
-    let result = result.as_ref().expect("result should be set");
-    assert!(result.success);
-    assert_eq!(result.exit_code, 0);
-    assert_eq!(result.stdout, "upgrade complete");
-    txn.commit().await.unwrap();
-
-    Ok(())
-}
-
-#[crate::sqlx_test]
-async fn test_report_scout_firmware_upgrade_status_failure(
-    pool: sqlx::PgPool,
-) -> CarbideResult<()> {
-    const UPGRADE_TASK_ID: &str = "scout-upgrade-task-id";
-
-    let env = create_test_env(pool).await;
-    let mh = common::api_fixtures::create_managed_host(&env).await;
-
-    // Manually put the machine into WaitingForScoutUpgrade state
-    let mut txn = env.pool.begin().await.unwrap();
-    let host = mh.host().db_machine(&mut txn).await;
-    let waiting_state = ManagedHostState::HostReprovision {
-        reprovision_state: HostReprovisionState::WaitingForScoutUpgrade {
-            upgrade_task_id: UPGRADE_TASK_ID.to_string(),
-            firmware_type: FirmwareComponentType::Bmc,
-            final_version: "1.2.3".to_string(),
-            power_drains_needed: None,
-            started_at: chrono::Utc::now(),
-            deadline: chrono::Utc::now() + chrono::TimeDelta::minutes(60),
-            task_json: String::new(),
-            result: None,
-        },
-        retry_count: 0,
-    };
-    db::machine::advance(&host, &mut txn, &waiting_state, None)
-        .await
-        .unwrap();
-    txn.commit().await.unwrap();
-
-    // Call the RPC endpoint with a failure result
-    env.api
-        .report_scout_firmware_upgrade_status(Request::new(
-            rpc::forge::ScoutFirmwareUpgradeStatusRequest {
-                machine_id: Some(mh.host().id),
-                success: false,
-                exit_code: 1,
-                stdout: "starting upgrade".to_string(),
-                stderr: "permission denied".to_string(),
-                error: "script failed".to_string(),
-                upgrade_task_id: UPGRADE_TASK_ID.to_string(),
-            },
-        ))
-        .await
-        .unwrap();
-
-    // Verify the failure result was stored
-    let mut txn = env.pool.begin().await.unwrap();
-    let host = mh.host().db_machine(&mut txn).await;
-    let ManagedHostState::HostReprovision {
-        reprovision_state, ..
-    } = host.current_state()
-    else {
-        panic!("Not in HostReprovision");
-    };
-    let HostReprovisionState::WaitingForScoutUpgrade { result, .. } = reprovision_state else {
-        panic!("Not in WaitingForScoutUpgrade");
-    };
-    let result = result.as_ref().expect("result should be set");
-    assert!(!result.success);
-    assert_eq!(result.exit_code, 1);
-    assert_eq!(result.stderr, "permission denied");
-    assert_eq!(result.error, "script failed");
-    txn.commit().await.unwrap();
-
-    Ok(())
-}
-
-#[crate::sqlx_test]
-async fn test_report_scout_firmware_upgrade_status_wrong_state(
-    pool: sqlx::PgPool,
-) -> CarbideResult<()> {
-    let env = create_test_env(pool).await;
-    let mh = common::api_fixtures::create_managed_host(&env).await;
-
-    // Machine is in its default state (not WaitingForScoutUpgrade), so the RPC should fail
-    let err = env
-        .api
-        .report_scout_firmware_upgrade_status(Request::new(
-            rpc::forge::ScoutFirmwareUpgradeStatusRequest {
-                machine_id: Some(mh.host().id),
-                upgrade_task_id: "scout-upgrade-task-id".to_string(),
-                success: true,
-                exit_code: 0,
-                stdout: String::new(),
-                stderr: String::new(),
-                error: String::new(),
-            },
-        ))
-        .await
-        .unwrap_err();
-
-    assert_eq!(err.code(), tonic::Code::FailedPrecondition);
-
-    Ok(())
-}
-
-#[crate::sqlx_test]
-async fn test_report_scout_firmware_upgrade_status_rejects_stale_task_id(
-    pool: sqlx::PgPool,
-) -> CarbideResult<()> {
-    const CURRENT_TASK_ID: &str = "current-scout-upgrade-task-id";
-    const STALE_TASK_ID: &str = "stale-scout-upgrade-task-id";
-
-    let env = create_test_env(pool).await;
-    let mh = common::api_fixtures::create_managed_host(&env).await;
-
-    let mut txn = env.pool.begin().await.unwrap();
-    let host = mh.host().db_machine(&mut txn).await;
-    let waiting_state = ManagedHostState::HostReprovision {
-        reprovision_state: HostReprovisionState::WaitingForScoutUpgrade {
-            upgrade_task_id: CURRENT_TASK_ID.to_string(),
-            firmware_type: FirmwareComponentType::Bmc,
-            final_version: "1.2.3".to_string(),
-            power_drains_needed: None,
-            started_at: chrono::Utc::now(),
-            deadline: chrono::Utc::now() + chrono::TimeDelta::minutes(60),
-            task_json: String::new(),
-            result: None,
-        },
-        retry_count: 0,
-    };
-    db::machine::advance(&host, &mut txn, &waiting_state, None)
-        .await
-        .unwrap();
-    txn.commit().await.unwrap();
-
-    let err = env
-        .api
-        .report_scout_firmware_upgrade_status(Request::new(
-            rpc::forge::ScoutFirmwareUpgradeStatusRequest {
-                machine_id: Some(mh.host().id),
-                success: true,
-                exit_code: 0,
-                stdout: "stale success".to_string(),
-                stderr: String::new(),
-                error: String::new(),
-                upgrade_task_id: STALE_TASK_ID.to_string(),
-            },
-        ))
-        .await
-        .unwrap_err();
-
-    assert_eq!(err.code(), tonic::Code::FailedPrecondition);
-
-    let mut txn = env.pool.begin().await.unwrap();
-    let host = mh.host().db_machine(&mut txn).await;
-    let ManagedHostState::HostReprovision {
-        reprovision_state, ..
-    } = host.current_state()
-    else {
-        panic!("Not in HostReprovision");
-    };
-    let HostReprovisionState::WaitingForScoutUpgrade {
-        upgrade_task_id,
-        result,
-        ..
-    } = reprovision_state
-    else {
-        panic!("Not in WaitingForScoutUpgrade");
-    };
-    assert_eq!(upgrade_task_id, CURRENT_TASK_ID);
-    assert!(result.is_none());
-    txn.commit().await.unwrap();
-
-    Ok(())
-}
-
-#[crate::sqlx_test]
-async fn test_report_scout_firmware_upgrade_status_truncates_output(
-    pool: sqlx::PgPool,
-) -> CarbideResult<()> {
-    const UPGRADE_TASK_ID: &str = "scout-upgrade-task-id";
-
-    let env = create_test_env(pool).await;
-    let mh = common::api_fixtures::create_managed_host(&env).await;
-
-    // Manually put the machine into WaitingForScoutUpgrade state
-    let mut txn = env.pool.begin().await.unwrap();
-    let host = mh.host().db_machine(&mut txn).await;
-    let waiting_state = ManagedHostState::HostReprovision {
-        reprovision_state: HostReprovisionState::WaitingForScoutUpgrade {
-            upgrade_task_id: UPGRADE_TASK_ID.to_string(),
-            firmware_type: FirmwareComponentType::Bmc,
-            final_version: "1.2.3".to_string(),
-            power_drains_needed: None,
-            started_at: chrono::Utc::now(),
-            deadline: chrono::Utc::now() + chrono::TimeDelta::minutes(60),
-            task_json: String::new(),
-            result: None,
-        },
-        retry_count: 0,
-    };
-    db::machine::advance(&host, &mut txn, &waiting_state, None)
-        .await
-        .unwrap();
-    txn.commit().await.unwrap();
-
-    // Send a response with very large stdout/stderr
-    let large_output = "x".repeat(10_000);
-    env.api
-        .report_scout_firmware_upgrade_status(Request::new(
-            rpc::forge::ScoutFirmwareUpgradeStatusRequest {
-                machine_id: Some(mh.host().id),
-                success: true,
-                exit_code: 0,
-                stdout: large_output.clone(),
-                stderr: large_output.clone(),
-                error: large_output.clone(),
-                upgrade_task_id: UPGRADE_TASK_ID.to_string(),
-            },
-        ))
-        .await
-        .unwrap();
-
-    // Verify the output was truncated
-    let mut txn = env.pool.begin().await.unwrap();
-    let host = mh.host().db_machine(&mut txn).await;
-    let ManagedHostState::HostReprovision {
-        reprovision_state, ..
-    } = host.current_state()
-    else {
-        panic!("Not in HostReprovision");
-    };
-    let HostReprovisionState::WaitingForScoutUpgrade { result, .. } = reprovision_state else {
-        panic!("Not in WaitingForScoutUpgrade");
-    };
-    let result = result.as_ref().expect("result should be set");
-    assert!(result.stdout.len() <= 1500);
-    assert!(result.stderr.len() <= 1500);
-    assert!(result.error.len() <= 1500);
-    txn.commit().await.unwrap();
-
-    Ok(())
-}
-
-/// Helper: set `host` to WaitingForScoutUpgrade with the given deadline and result.
-async fn put_in_waiting_for_scout_upgrade(
-    env: &common::api_fixtures::TestEnv,
-    host: &common::api_fixtures::test_machine::TestMachine,
-    deadline: chrono::DateTime<chrono::Utc>,
-    power_drains_needed: Option<u32>,
-    result: Option<model::machine::ScoutUpgradeResult>,
-) {
-    let mut txn = env.pool.begin().await.unwrap();
-    let machine = host.db_machine(&mut txn).await;
-    let state = ManagedHostState::HostReprovision {
-        reprovision_state: HostReprovisionState::WaitingForScoutUpgrade {
-            upgrade_task_id: "scout-upgrade-task-id".to_string(),
-            firmware_type: FirmwareComponentType::Bmc,
-            final_version: "1.2.3".to_string(),
-            power_drains_needed,
-            started_at: chrono::Utc::now(),
-            deadline,
-            task_json: String::new(),
-            result,
-        },
-        retry_count: 0,
-    };
-    db::machine::advance(&machine, &mut txn, &state, None)
-        .await
-        .unwrap();
-    txn.commit().await.unwrap();
-}
-
-async fn put_in_new_firmware_reported_wait(
-    env: &TestEnv,
-    mh: &TestManagedHost,
-    previous_reset_time: i64,
-    reset_retry_count: u32,
-) {
-    let mut txn = env.pool.begin().await.unwrap();
-    let machine = mh.host().db_machine(&mut txn).await;
-    let state = ManagedHostState::HostReprovision {
-        reprovision_state: HostReprovisionState::NewFirmwareReportedWait {
-            firmware_type: FirmwareComponentType::Uefi,
-            firmware_number: Some(0),
-            final_version: "1.13.2".to_string(),
-            previous_reset_time: Some(previous_reset_time),
-            reset_retry_count,
-        },
-        retry_count: 0,
-    };
-    db::machine::advance(&machine, &mut txn, &state, None)
-        .await
-        .unwrap();
-    txn.commit().await.unwrap();
-}
-
-#[crate::sqlx_test]
-async fn test_new_firmware_reported_wait_retries_reset_after_timeout(
-    pool: sqlx::PgPool,
-) -> CarbideResult<()> {
-    let env = create_test_env(pool).await;
-    let mh = common::api_fixtures::create_managed_host(&env).await;
-    put_in_new_firmware_reported_wait(&env, &mh, chrono::Utc::now().timestamp() - 31 * 60, 0).await;
-
-    env.run_machine_state_controller_iteration().await;
-
-    let mut txn = env.pool.begin().await.unwrap();
-    let host = mh.host().db_machine(&mut txn).await;
-    let ManagedHostState::HostReprovision {
-        reprovision_state, ..
-    } = host.current_state()
-    else {
-        panic!("Not in HostReprovision");
-    };
-    let HostReprovisionState::NewFirmwareReportedWait {
-        reset_retry_count, ..
-    } = reprovision_state
-    else {
-        panic!("expected NewFirmwareReportedWait, got {reprovision_state:?}");
-    };
-    assert_eq!(*reset_retry_count, 1);
-
-    Ok(())
-}
-
-#[crate::sqlx_test]
-async fn test_new_firmware_reported_wait_fails_after_reset_retry_limit(
-    pool: sqlx::PgPool,
-) -> CarbideResult<()> {
-    let env = create_test_env(pool).await;
-    let mh = common::api_fixtures::create_managed_host(&env).await;
-    put_in_new_firmware_reported_wait(
-        &env,
-        &mh,
-        chrono::Utc::now().timestamp() - 31 * 60,
-        MAX_NEW_FIRMWARE_REPORTED_RESET_RETRIES,
-    )
-    .await;
-
-    env.run_machine_state_controller_iteration().await;
-
-    let mut txn = env.pool.begin().await.unwrap();
-    let host = mh.host().db_machine(&mut txn).await;
-    let ManagedHostState::HostReprovision {
-        reprovision_state, ..
-    } = host.current_state()
-    else {
-        panic!("Not in HostReprovision");
-    };
-    let HostReprovisionState::FailedFirmwareUpgrade { reason, .. } = reprovision_state else {
-        panic!("expected FailedFirmwareUpgrade, got {reprovision_state:?}");
-    };
-    let reason = reason.as_deref().unwrap_or_default();
-    assert!(
-        reason.contains("Firmware version did not converge after completed update"),
-        "unexpected reason: {reason}",
-    );
-    assert!(
-        reason.contains("expected 1.13.2"),
-        "unexpected reason: {reason}",
-    );
-    assert!(reason.contains("1.12.0"), "unexpected reason: {reason}",);
-
-    Ok(())
-}
-
-#[crate::sqlx_test]
-async fn test_waiting_for_scout_upgrade_success_transitions_to_reset_for_new_firmware(
-    pool: sqlx::PgPool,
-) -> CarbideResult<()> {
-    let env = create_test_env(pool).await;
-    let mh = common::api_fixtures::create_managed_host(&env).await;
-
-    put_in_waiting_for_scout_upgrade(
-        &env,
-        &mh.host(),
-        chrono::Utc::now() + chrono::TimeDelta::minutes(60),
-        Some(1),
-        Some(model::machine::ScoutUpgradeResult {
-            success: true,
-            exit_code: 0,
-            stdout: "ok".to_string(),
-            stderr: String::new(),
-            error: String::new(),
-        }),
-    )
-    .await;
-
-    env.run_machine_state_controller_iteration().await;
-
-    let mut txn = env.pool.begin().await.unwrap();
-    let host = mh.host().db_machine(&mut txn).await;
-    let ManagedHostState::HostReprovision {
-        reprovision_state, ..
-    } = host.current_state()
-    else {
-        panic!("Not in HostReprovision");
-    };
-    let HostReprovisionState::ResetForNewFirmware {
-        power_drains_needed,
-        ..
-    } = reprovision_state
-    else {
-        panic!("expected ResetForNewFirmware, got {reprovision_state:?}");
-    };
-    assert_eq!(*power_drains_needed, Some(1));
-
-    Ok(())
-}
-
-#[crate::sqlx_test]
-async fn test_waiting_for_scout_upgrade_failure_uses_error_as_reason(
-    pool: sqlx::PgPool,
-) -> CarbideResult<()> {
-    let env = create_test_env(pool).await;
-    let mh = common::api_fixtures::create_managed_host(&env).await;
-
-    put_in_waiting_for_scout_upgrade(
-        &env,
-        &mh.host(),
-        chrono::Utc::now() + chrono::TimeDelta::minutes(60),
-        None,
-        Some(model::machine::ScoutUpgradeResult {
-            success: false,
-            exit_code: 1,
-            stdout: String::new(),
-            stderr: String::new(),
-            error: "boom".to_string(),
-        }),
-    )
-    .await;
-
-    env.run_machine_state_controller_iteration().await;
-
-    let mut txn = env.pool.begin().await.unwrap();
-    let host = mh.host().db_machine(&mut txn).await;
-    let ManagedHostState::HostReprovision {
-        reprovision_state, ..
-    } = host.current_state()
-    else {
-        panic!("Not in HostReprovision");
-    };
-    let HostReprovisionState::FailedFirmwareUpgrade { reason, .. } = reprovision_state else {
-        panic!("expected FailedFirmwareUpgrade, got {reprovision_state:?}");
-    };
-    assert_eq!(reason.as_deref(), Some("boom"));
-
-    Ok(())
-}
-
-#[crate::sqlx_test]
-async fn test_waiting_for_scout_upgrade_failure_without_error_uses_exit_code(
-    pool: sqlx::PgPool,
-) -> CarbideResult<()> {
-    let env = create_test_env(pool).await;
-    let mh = common::api_fixtures::create_managed_host(&env).await;
-
-    put_in_waiting_for_scout_upgrade(
-        &env,
-        &mh.host(),
-        chrono::Utc::now() + chrono::TimeDelta::minutes(60),
-        None,
-        Some(model::machine::ScoutUpgradeResult {
-            success: false,
-            exit_code: 7,
-            stdout: String::new(),
-            stderr: String::new(),
-            error: String::new(),
-        }),
-    )
-    .await;
-
-    env.run_machine_state_controller_iteration().await;
-
-    let mut txn = env.pool.begin().await.unwrap();
-    let host = mh.host().db_machine(&mut txn).await;
-    let ManagedHostState::HostReprovision {
-        reprovision_state, ..
-    } = host.current_state()
-    else {
-        panic!("Not in HostReprovision");
-    };
-    let HostReprovisionState::FailedFirmwareUpgrade { reason, .. } = reprovision_state else {
-        panic!("expected FailedFirmwareUpgrade, got {reprovision_state:?}");
-    };
-    assert_eq!(
-        reason.as_deref(),
-        Some("Scout upgrade failed with exit code 7"),
-    );
-
-    Ok(())
-}
-
-#[crate::sqlx_test]
-async fn test_waiting_for_scout_upgrade_past_deadline_times_out(
-    pool: sqlx::PgPool,
-) -> CarbideResult<()> {
-    let env = create_test_env(pool).await;
-    let mh = common::api_fixtures::create_managed_host(&env).await;
-
-    put_in_waiting_for_scout_upgrade(
-        &env,
-        &mh.host(),
-        chrono::Utc::now() - chrono::TimeDelta::minutes(1),
-        None,
-        None,
-    )
-    .await;
-
-    env.run_machine_state_controller_iteration().await;
-
-    let mut txn = env.pool.begin().await.unwrap();
-    let host = mh.host().db_machine(&mut txn).await;
-    let ManagedHostState::HostReprovision {
-        reprovision_state, ..
-    } = host.current_state()
-    else {
-        panic!("Not in HostReprovision");
-    };
-    let HostReprovisionState::FailedFirmwareUpgrade { reason, .. } = reprovision_state else {
-        panic!("expected FailedFirmwareUpgrade, got {reprovision_state:?}");
-    };
-    assert!(
-        reason
-            .as_deref()
-            .is_some_and(|r| r.starts_with("Scout firmware upgrade timed out")),
-        "unexpected reason: {reason:?}",
-    );
-
-    Ok(())
-}
-
-#[crate::sqlx_test]
-async fn test_waiting_for_scout_upgrade_before_deadline_waits(
-    pool: sqlx::PgPool,
-) -> CarbideResult<()> {
-    let env = create_test_env(pool).await;
-    let mh = common::api_fixtures::create_managed_host(&env).await;
-
-    put_in_waiting_for_scout_upgrade(
-        &env,
-        &mh.host(),
-        chrono::Utc::now() + chrono::TimeDelta::minutes(60),
-        None,
-        None,
-    )
-    .await;
-
-    env.run_machine_state_controller_iteration().await;
-
-    let mut txn = env.pool.begin().await.unwrap();
-    let host = mh.host().db_machine(&mut txn).await;
-    let ManagedHostState::HostReprovision {
-        reprovision_state, ..
-    } = host.current_state()
-    else {
-        panic!("Not in HostReprovision");
-    };
-    assert!(
-        matches!(
-            reprovision_state,
-            HostReprovisionState::WaitingForScoutUpgrade { result: None, .. }
-        ),
-        "expected to remain in WaitingForScoutUpgrade, got {reprovision_state:?}",
-    );
 
     Ok(())
 }
